Sequential learning not quite working
