+++ dependencies.txt
@@ -7,3 +7,4 @@ keras==2.1.5
 h5py==2.7.1
 pillow==5.0.0
 sklearn==0.0
+IPython==5.7.0
