+++ src/learning/model.py
@@ -27,8 +27,8 @@ IMAGE_SIZE = IMAGE_WIDTH * IMAGE_HEIGHT
 COLOR_CHANNELS = 3
 WEIGHT_SIZE = 5
 BATCH_SIZE = 50  # Number of images per step or iteration
-KEEP_RATE = 0.85
-N_EPOCHS = 10  # One iteration over all of the training data
+KEEP_RATE = 0.55
+N_EPOCHS = 3500  # One iteration over all of the training data
 FC_NEURON_SIZE = 1024  # Chosen randomly for now
 N_CLASSES = len(CLASSES)
 FC_NUM_FEATURES = np.int32(IMAGE_WIDTH * IMAGE_HEIGHT * N_CLASSES * .8)
